Separate package prepare script from npm publish script
