+++ package.json
@@ -48,7 +48,7 @@
     "lint:js": "eslint --ext .js,.ts,.tsx src",
     "now-build": "npm run build:storybook",
     "prepare:docs": "npm run prepare:package && npm run copy-changelog",
-    "prepare:package": "npm run build && rm -rf pkg && mkdir pkg && cp -r dist pkg && cp package.json pkg/package.json",
+    "prepare:package": "npm run build && node scripts/prepublish",
     "publish": "npm run prepare:package && node scripts/publish",
     "storybook": "start-storybook -p 6006",
     "test": "stencil test --spec --e2e --maxWorkers=2",
